% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_catalogue}
\alias{assert_catalogue}
\title{Assert object represents a signature}
\usage{
assert_catalogue(
  obj,
  must_sum_to_one = TRUE,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{obj}{object}

\item{must_sum_to_one}{throw an error if the fraction column of the signature data.frame does NOT sum to one.}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
Throws error if assertion fails, otherwise invisibly returns true
}
\description{
Assert object represents a catalogue
}
\examples{
# Generate Example Datatypes

if(interactive()){

  # Signatures
  signature = example_valid_signature()
  signature_collection = example_valid_signature_collection()
  signature_annotations = example_valid_signature_annotations()

  # Catalogues
  catalogue = example_valid_catalogue()
  catalogue_collection = example_valid_catalogue_collection()

  # Cohort Analysis Results
  cohort_analysis = example_valid_cohort_analysis()

  # Cohort Analysis Results
  catalogue_collection = example_valid_catalogue_collection()

  # Run Assertions
  assert_signature(signature)
  assert_signature_collection(signature_collection)
  assert_catalogue(catalogue)
  assert_catalogue_collection(catalogue_collection)
  assert_cohort_analysis(cohort_analysis)
}
}
