% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_signature_collection}
\alias{assert_signature_collection}
\title{Assert object represents a signature}
\usage{
assert_signature_collection(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
Throws error if assertion fails, otherwise invisibly returns true
}
\description{
Assert object represents a signature collection
}
\examples{
# Generate Example Datatypes

if(interactive()){

  # Signatures
  signature = example_valid_signature()
  signature_collection = example_valid_signature_collection()
  signature_annotations = example_valid_signature_annotations()

  # Decompositions
  decomposition = example_valid_decomposition()
  decomposition_collection = example_valid_decomposition_collection()

  # Cohort Analysis Results
  cohort_analysis = example_valid_cohort_analysis()

  # Cohort Analysis Results
  decomposition_collection = example_valid_decomposition_collection()

  # Run Assertions
  assert_signature(signature)
  assert_signature_collection(signature_collection)
  assert_decomposition(decomposition)
  assert_decomposition_collection(decomposition_collection)
  assert_cohort_analysis(cohort_analysis)
}
}
