% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_bootstraps}
\alias{assert_bootstraps}
\alias{bootstraps}
\alias{example_bootstraps}
\title{Bootstraps Data Type}
\usage{
assert_bootstraps(obj, msg = NULL, call = rlang::caller_env(), arg_name = NULL)

example_bootstraps()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A data.frame representing a valid exemplar bootstrap in the 'sigverse' format.
}
\description{
Assert object represents a bootstrap data.frame.

\strong{Bootstraps}: A \code{data.frame} representing the optimal mutational signature exposure contributions for each bootstrap resample. Must contain 1 row per signature & bootstrap combination with the following 5 columns:
\itemize{
\item \code{bootstrap}: The bootstrap index.
\item \code{signature}: The name of the signature.
\item \code{contribution_absolute}: The absolute contribution of the signature.
\item \code{contribution}: The percentage contribution of the signature.
}

This function returns an exemplar bootstrap with valid data following the 'sigverse' style.
}
\examples{

# Return a valid bootstrap
example_bootstraps()

# Return a valid but empty bootstrap (all contributions are 0)
example_bootstraps_empty()
}
