% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_catalogue}
\alias{assert_catalogue}
\alias{catalogue}
\alias{example_catalogue}
\title{Catalogue Data Type}
\usage{
assert_catalogue(
  obj,
  must_sum_to_one = TRUE,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_catalogue()
}
\arguments{
\item{obj}{object}

\item{must_sum_to_one}{throw an error if the fraction column of the signature data.frame does NOT sum to one.}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A data.frame containing valid catalogue data.
}
\description{
Assert object represents a catalogue

\strong{Catalogue}: The mutational profile of a sample, represented as a \code{data.frame} with 4 required columns:
\itemize{
\item \code{channel}: The mutational channel (e.g., A\[T>C\]G, C\[C>T\]G).
\item \code{type}: The type of mutation (higher level classification of channel, e.g. T>C, C>T).
\item \code{fraction}: The fraction of mutations attributed to this channel.
\item \code{count}: The count of mutations for this channel.
}

This function returns a data.frame representing a valid catalogue following the 'sigverse' style.
}
\examples{

# Example of a valid catalogue
example_catalogue()

# Example of a valid but empty catalogue
example_catalogue_empty()

}
