% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_catalogue_collection}
\alias{assert_catalogue_collection}
\alias{catalogue_collections}
\alias{example_catalogue_collection}
\title{Catalogue Collections Data Type}
\usage{
assert_catalogue_collection(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_catalogue_collection()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A list containing multiple data.frames, each representing a valid catalogue.
}
\description{
Assert object represents a collection of catalogues

\strong{Catalogue Collections}: A list of catalogue \code{data.frames}, where each \code{data.frame} represents the mutational profile of a sample. Each entry in the list corresponds to a sample, and the name of each entry is the sample identifier. Each catalogue \code{data.frame} contains the following columns:
\itemize{
\item \code{channel}: The mutational channel (e.g., A[T>C]G, C[C>T]G).
\item \code{type}: The type of mutation (higher level classification of channel, e.g. T>C, C>T).
\item \code{fraction}: The fraction of mutations attributed to this channel.
\item \code{count}: The count of mutations for this channel.
}

This function returns a list of exemplar catalogues with valid data following the 'sigverse' style. The list represents a collection of valid catalogues.
}
\examples{
example_catalogue_collection()

}
