% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_cohort_analysis}
\alias{assert_cohort_analysis}
\alias{cohort}
\alias{example_cohort_analysis}
\title{Cohort Signature Analysis Results Data Type}
\usage{
assert_cohort_analysis(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_cohort_analysis()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A data.frame containing multiple cohort analyses.
}
\description{
Assert object represents results of a cohort analysis

\strong{Cohort}: A \code{data.frame} containing results from a cohort-level signature analysis, with the following 5 columns:
\itemize{
\item \code{sample}: The sample identifier.
\item \code{signature}: The name of the signature.
\item \code{contribution_absolute}: The absolute contribution of the signature to the sample.
\item \code{contribution}: The proportional contribution of the signature.
\item \code{p_value}: The p-value representing the statistical significance of the contribution.
}

This function returns a data.frame representing a collection of valid cohort analyses following the 'sigverse' style.
}
\examples{
example_cohort_analysis()

}
