% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/assertions.R
\name{example_cohort_metadata}
\alias{example_cohort_metadata}
\alias{assert_cohort_metadata}
\alias{cohort_metadata}
\title{Sigverse Cohort Metadata}
\usage{
example_cohort_metadata()

assert_cohort_metadata(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
\code{\link[=example_cohort_metadata]{example_cohort_metadata()}} returns an exemplar data.frame describing minimal sample-level metadata.

\code{\link[=assert_cohort_metadata]{assert_cohort_metadata()}} throws error (if assertion fails) or invisibly returns TRUE if successful.
}
\description{
\strong{Cohort Metadata}: A \code{data.frame} describing sample-level metadata. Requires the following columns:
\itemize{
\item \code{sample}: The sample identifier.
\item \code{disease}: The disease from which this sample arose (e.g. Lung Cancer / Melanoma).
}

Can also include any number of additional columns with other metadata.

\code{\link[=example_cohort_metadata]{example_cohort_metadata()}} returns an exemplar 'sigverse' style cohort metadata data.frame

\code{\link[=assert_cohort_metadata]{assert_cohort_metadata()}} asserts an object is a valid sigverse-style cohort metadata data.frame
}
\examples{

# Example sample level metadata for a cohort
metadata <- example_cohort_metadata()

# Assert that metadata contains all the columns the sigverse requires
assert_cohort_metadata(metadata)

}
