% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_annotations}
\alias{example_annotations}
\title{Exemplar collection of valid annotations}
\usage{
example_annotations()
}
\value{
A data.frame containing annotations for multiple signatures.
}
\description{
This function returns a data.frame representing a collection of valid annotations following the 'sigverse' style.
}
\examples{
example_annotations()

}
