% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_bootstraps_colo829}
\alias{example_bootstraps_colo829}
\title{Exemplar bootstraps with valid data}
\usage{
example_bootstraps_colo829()
}
\value{
A data.frame representing bootstraps from SBS mutational signature analysis of COLO829
in the 'sigverse' format.
}
\description{
This function returns example bootstrap data from somatic Single Base Substitution (SBS)
mutational signature analysis of the COLO829 melanoma cell line, following the 'sigverse' format.
}
\examples{

# Return example bootstraps from COLO829 SBS analysis
example_bootstraps_colo829()

}
