% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_catalogue_colo829}
\alias{example_catalogue_colo829}
\title{Exemplar catalogue with valid data}
\usage{
example_catalogue_colo829()
}
\value{
A data.frame representing the SBS mutational catalogue of COLO829 in the 'sigverse' format
(see \code{\link[=example_catalogue]{example_catalogue()}}).
}
\description{
This function returns example catalogue from somatic Single Base Substitution (SBS)
mutational signature analysis of the COLO829 melanoma cell line, following the 'sigverse' format
(see \code{\link[=example_catalogue]{example_catalogue()}}).
}
\examples{

# Return example catalogue from COLO829 SBS analysis
example_catalogue_colo829()

}
