% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_cohort_metadata_empty}
\alias{example_cohort_metadata_empty}
\title{Example Data}
\usage{
example_cohort_metadata_empty()
}
\description{
See \code{\link[=example_cohort_metadata]{example_cohort_metadata()}}
}
