% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_bootstrap_contribution}
\alias{example_invalid_bootstrap_contribution}
\title{Invalid bootstrap with contribution > 100}
\usage{
example_invalid_bootstrap_contribution()
}
\value{
A data.frame representing an invalid bootstrap with contribution > 1
}
\description{
This function returns a bootstrap with invalid data where contribution percentage is greater than 100.
}
