% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_bootstrap_missing}
\alias{example_invalid_bootstrap_missing}
\title{Invalid bootstrap with missing data}
\usage{
example_invalid_bootstrap_missing()
}
\value{
A data.frame representing an invalid bootstrap with missing data.
}
\description{
This function returns a bootstrap with missing data (NA values).
}
