% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_bootstraps_negative_contribution}
\alias{example_invalid_bootstraps_negative_contribution}
\title{Invalid bootstrap with negative contribution}
\usage{
example_invalid_bootstraps_negative_contribution()
}
\value{
A data.frame representing an invalid bootstrap with negative contribution.
}
\description{
This function returns a bootstrap with invalid data where contribution values are negative.
}
