% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_bootstraps_negative_contribution_absolute}
\alias{example_invalid_bootstraps_negative_contribution_absolute}
\title{Invalid bootstrap with negative contribution_absolute}
\usage{
example_invalid_bootstraps_negative_contribution_absolute()
}
\value{
A data.frame representing an invalid bootstrap with negative contribution_absolute.
}
\description{
This function returns a bootstrap with invalid data where absolute contribution values are negative.
}
