% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_similarity_missing_cosine}
\alias{example_invalid_similarity_missing_cosine}
\title{Exemplar \verb{Similarity against cohort dataset} with missing 'cosine_similarity' column}
\usage{
example_invalid_similarity_missing_cosine()
}
\value{
A \code{data.frame} missing the 'cosine_similarity' column.
}
\description{
Exemplar \verb{Similarity against cohort dataset} with missing 'cosine_similarity' column
}
