% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_similarity_na_in_cosine}
\alias{example_invalid_similarity_na_in_cosine}
\title{Exemplar \verb{Similarity against cohort dataset} with NA in 'cosine_similarity' column}
\usage{
example_invalid_similarity_na_in_cosine()
}
\value{
A \code{data.frame} with \code{NA} values in 'cosine_similarity' column.
}
\description{
Exemplar \verb{Similarity against cohort dataset} with NA in 'cosine_similarity' column
}
