% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_similarity_non_character_sample}
\alias{example_invalid_similarity_non_character_sample}
\title{Exemplar \verb{Similarity against cohort dataset} with non-character 'sample' column}
\usage{
example_invalid_similarity_non_character_sample()
}
\value{
A \code{data.frame} where 'sample' is not character or factor.
}
\description{
Exemplar \verb{Similarity against cohort dataset} with non-character 'sample' column
}
