% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_similarity_non_numeric_cosine}
\alias{example_invalid_similarity_non_numeric_cosine}
\title{Exemplar \verb{Similarity against cohort dataset} with non-numeric 'cosine_similarity' column}
\usage{
example_invalid_similarity_non_numeric_cosine()
}
\value{
A \code{data.frame} where 'cosine_similarity' is not numeric.
}
\description{
Exemplar \verb{Similarity against cohort dataset} with non-numeric 'cosine_similarity' column
}
