% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_umap_duplicate_samples}
\alias{example_invalid_umap_duplicate_samples}
\title{Exemplar UMAP dataset with duplicated samples}
\usage{
example_invalid_umap_duplicate_samples()
}
\value{
A \code{data.frame} with duplicated sample identifiers.
}
\description{
Exemplar UMAP dataset with duplicated samples
}
