% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_umap_missing_dim1}
\alias{example_invalid_umap_missing_dim1}
\title{Exemplar UMAP dataset with missing 'dim1' column}
\usage{
example_invalid_umap_missing_dim1()
}
\value{
A \code{data.frame} missing the 'dim1' column.
}
\description{
Exemplar UMAP dataset with missing 'dim1' column
}
