% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_umap_missing_sample}
\alias{example_invalid_umap_missing_sample}
\title{Exemplar UMAP dataset with missing 'sample' column}
\usage{
example_invalid_umap_missing_sample()
}
\value{
A \code{data.frame} missing the 'sample' column.
}
\description{
Exemplar UMAP dataset with missing 'sample' column
}
