% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_umap_na_in_sample}
\alias{example_invalid_umap_na_in_sample}
\title{Exemplar UMAP dataset with missing values in 'sample' column}
\usage{
example_invalid_umap_na_in_sample()
}
\value{
A \code{data.frame} with \code{NA} values in 'sample' column.
}
\description{
Exemplar UMAP dataset with missing values in 'sample' column
}
