% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_invalid_umap_non_numeric_dim1}
\alias{example_invalid_umap_non_numeric_dim1}
\title{Exemplar UMAP dataset with non-numeric 'dim1' column}
\usage{
example_invalid_umap_non_numeric_dim1()
}
\value{
A \code{data.frame} where 'dim1' is not numeric.
}
\description{
Exemplar UMAP dataset with non-numeric 'dim1' column
}
