% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_catalogue}
\alias{example_valid_catalogue}
\title{Exemplar valid catalogue}
\usage{
example_valid_catalogue()
}
\value{
A data.frame containing valid catalogue data.
}
\description{
This function returns a data.frame representing a valid catalogue following the 'sigverse' style.
}
\examples{
example_valid_catalogue()

}
