% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_catalogue_collection}
\alias{example_valid_catalogue_collection}
\title{Exemplar collection of valid catalogues}
\usage{
example_valid_catalogue_collection()
}
\value{
A list containing multiple data.frames, each representing a valid catalogue.
}
\description{
This function returns a list of exemplar catalogues with valid data following the 'sigverse' style. The list represents a collection of valid catalogues.
}
\examples{
example_valid_catalogue_collection()

}
