% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_cohort_analysis}
\alias{example_valid_cohort_analysis}
\title{Exemplar collection of valid cohort analyses}
\usage{
example_valid_cohort_analysis()
}
\value{
A data.frame containing multiple cohort analyses.
}
\description{
This function returns a data.frame representing a collection of valid cohort analyses following the 'sigverse' style.
}
\examples{
example_valid_cohort_analysis()

}
