% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_decomposition}
\alias{example_valid_decomposition}
\title{Exemplar valid decomposition}
\usage{
example_valid_decomposition()
}
\value{
A data.frame containing valid decomposition data.
}
\description{
This function returns a data.frame representing a valid decomposition following the 'sigverse' style.
}
\examples{
example_valid_decomposition()

}
