% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_decomposition_collection}
\alias{example_valid_decomposition_collection}
\title{Exemplar collection of valid decompositions}
\usage{
example_valid_decomposition_collection()
}
\value{
A list containing multiple data.frames, each representing a valid decomposition.
}
\description{
This function returns a list of exemplar decompositions with valid data following the 'sigverse' style. The list represents a collection of valid decompositions.
}
\examples{
example_valid_decomposition_collection()

}
