% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_signature}
\alias{example_valid_signature}
\title{Exemplar signature with valid data}
\usage{
example_valid_signature()
}
\value{
A data.frame representing a valid signature.
}
\description{
This function returns an exemplar signature with valid data following the 'sigverse' style.
}
\examples{
example_valid_signature()

}
