% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_valid_signature_collection}
\alias{example_valid_signature_collection}
\title{Exemplar collection of valid signatures}
\usage{
example_valid_signature_collection()
}
\value{
A list containing multiple data.frames, each representing a valid signature.
}
\description{
This function returns a list of exemplar signatures with valid data following the 'sigverse' style. The list represents a collection of valid signatures.
}
\examples{
example_valid_signature_collection()

}
