% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/assertions.R
\name{example_model}
\alias{example_model}
\alias{assert_model}
\alias{model}
\title{Exemplar model specification}
\usage{
example_model()

assert_model(
  obj,
  signature_collection = NULL,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)
}
\arguments{
\item{obj}{object}

\item{signature_collection}{Optionally assert all signatures described in model are present a specific signature collection.}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
\code{\link[=example_model]{example_model()}} returns a named numeric vector where names correspond
signatures and matched values represents their proportional contribution.
}
\description{
\strong{Model}: A named numeric vector representing a signature model where names are signatures and the values are their proportional contributions to the model.

\code{\link[=example_model]{example_model()}} returns an exemplar 'sigverse' style signature model specification.
The model is represented as a named numeric vector where each name corresponds
to a signature, and each value represents its proportional contribution.

\code{\link[=assert_model]{assert_model()}} asserts an object is a valid sigverse model.
}
\examples{
# Return a valid signature model (representing a combination of
# 30\% Signature1 & 70\% Signature2)
example_model()

# Return a valid but empty signature model (length zero)
example_model_empty()

# Assert model is valid
assert_model(example_model())

# Assert all signatures described by model are in a signature collection
assert_model(example_model(), example_signature_collection())

}
