% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylikes.R
\name{rename}
\alias{rename}
\title{Rename columns in a data frame based on a name mapping}
\usage{
rename(.data, namemap)
}
\arguments{
\item{namemap}{A named vector or list, where the names represent the new column
names and the values represent the current column names in the data frame.}

\item{df}{A data frame whose columns will be renamed.}
}
\value{
The data frame with renamed columns.
}
\description{
This function renames columns in a data frame using a provided name mapping.
It verifies that all names in the mapping exist in the data frame before
renaming, and raises an error if any expected columns are missing.
}
\details{
The function checks if all the values (current column names) in the
\code{namemap} are present in the data frame. If any expected column names are missing,
the function stops with an error. If all expected names are present, the corresponding
columns are renamed according to the names provided in \code{namemap}.
}
\examples{
rename(mtcars, c(miles_per_gallon = "mpg"))
}
