% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_classes.R
\name{sig_aetiology_classes}
\alias{sig_aetiology_classes}
\title{Signature Aetiology Data-Dictionary}
\usage{
sig_aetiology_classes()
}
\value{
a data.frame describing valid terms to describe signature aetiology class/subclass
}
\description{
Signature Aetiology Data-Dictionary
}
\examples{
sig_aetiology_classes()
}
