% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_collection_reformat_list_to_matrix}
\alias{sig_collection_reformat_list_to_matrix}
\title{Convert Signature or Catalogue Collection to Matrix}
\usage{
sig_collection_reformat_list_to_matrix(
  signatures,
  values = c("fraction", "count")
)
}
\arguments{
\item{signatures}{A sigverse signature or catalogue collection. A named list of sigstash signature data.frames.
See \code{\link[=example_signature_collection]{example_signature_collection()}} or \code{\link[=example_catalogue_collection]{example_catalogue_collection()}}.}

\item{values}{Either \code{"fraction"} (default) or \code{"count"}. Use \code{"count"} only with catalogue collections.}
}
\value{
A numeric matrix with:
\itemize{
\item \strong{Rows}: Channels
\item \strong{Columns}: Signature names
\item \strong{Values}: Fractions or counts (based on \code{values})
}

Includes a \code{types} attribute mapping mutation types to their corresponding channels.
}
\description{
Transforms a signature or catalogue collection into a matrix format,
with channels as rows and signatures as columns.
}
\details{
The resulting matrix includes an attribute \code{type}. This vector describing mutation type of each row (channel) in matrix.
}
\examples{
# Convert a signature collection
collection <- example_signature_collection()
mat <- sig_collection_reformat_list_to_matrix(collection)

# Convert a catalogue collection using counts
catalogue <- example_catalogue_collection()
mat_counts <- sig_collection_reformat_list_to_matrix(catalogue, values = "count")
}
