% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_signature_analysis_results.R
\name{signature_analysis_result}
\alias{signature_analysis_result}
\title{Create a signature_analysis_result Object}
\usage{
signature_analysis_result(
  sample,
  sigclass,
  model,
  number_of_mutations,
  model_fit,
  unexplained_mutations,
  catalogue,
  signatures,
  signature_annotations,
  analysis_details,
  bootstraps,
  cohort_exposures = NULL,
  similarity_against_cohort = NULL,
  max_similar_samples = NULL,
  cohort_size = NULL,
  cohort_catalogues = NULL,
  cohort_metadata = NULL,
  umap = NULL
)
}
\arguments{
\item{sample}{String. Sample Identifier.}

\item{sigclass}{String. The class of signature features used (e.g., SBS96, ID83, SV32, etc).}

\item{model}{A named numeric vector where names are signatures and values are their proportional contributions to the model.}

\item{number_of_mutations}{Number. Total number of mutations in this sigclass.}

\item{model_fit}{Number. Measured as cosine similarity between observed profile vs model.}

\item{unexplained_mutations}{Number. Number of mutations not explained by signature.}

\item{catalogue}{data.frame. Observed mutation catalogue of the sample. See \link{example_catalogue}.}

\item{signatures}{The signature collection used for fitting, in sigstash format (see \code{\link[=example_signature_collection]{example_signature_collection()}} and \code{\link[sigstash:sig_load]{sigstash::sig_load()}}). Instead of supplying full signature dataset, can also be a string descibing a dataset name recognised by \code{\link[sigstash:sig_load]{sigstash::sig_load()}}.}

\item{signature_annotations}{Signature annotations describing aetiology for signatures in signatures (see \code{\link[=example_annotations]{example_annotations()}} and \code{\link[sigstash:sig_load_annotations]{sigstash::sig_load_annotations()}}). Instead of supplying full signature annotation dataset, can also be a string descibing a dataset name recognised by \code{\link[sigstash:sig_load_annotations]{sigstash::sig_load_annotations()}}.}

\item{analysis_details}{data.frame. describing the details of the analysis (software versions, thresholds, reference cohorts, etc).}

\item{bootstraps}{A dataframe in sigverse style describing bootstraps. See \code{\link[sigshared:bootstraps]{sigshared::example_bootstraps()}}.}

\item{cohort_exposures}{data.frame. A cohort of signature analysis results used to produce cohort distribution dotplots. See \code{\link[=example_cohort_analysis]{example_cohort_analysis()}}}

\item{similarity_against_cohort}{data.frame describing how similar the observed mutational catalogue is to other samples. See \code{\link[=example_similarity_against_cohort]{example_similarity_against_cohort()}}.
If including only a subset of the N most similar samples, please supply the total number of comparison samples to the \code{cohort_size} argument.}

\item{max_similar_samples}{If \code{similarity_against_cohort} describes only the N most similar samples, please set max_similar_samples to the value of N.
If \code{similarity_against_cohort} describes the full reference dataset, set to Inf.}

\item{cohort_size}{Number. How many samples are in the reference cohort this sample was compared to (not including self).}

\item{cohort_catalogues}{a sigverse collection of catalogues in the reference-cohort which the sample was compared against. Must at least include a decomposition for all samples in similarity_against_cohort. Used to create similar sample plots. See \code{\link[=example_catalogue_collection]{example_catalogue_collection()}} for example format.}

\item{cohort_metadata}{data.frame. Sample level metadata describing every sample in \code{cohort_catalogues} See \code{\link[=example_cohort_metadata]{example_cohort_metadata()}}.}

\item{umap}{data.frame. UMAP datasetframe showing how the catalogue of the sample-of-interest clusters against cohort_metadata. See \code{\link[=example_umap]{example_umap()}}.}
}
\value{
An object of class \code{sigstory} containing the provided data and visualizations.
}
\description{
Constructs a \code{signature_analysis_result} object containing all the information required to generate a \code{\link[=sigstory_visualisation]{sigstory_visualisation()}} object.
}
\note{
Many of the visualizations required for this object can be produced using the \code{sigvis} package.
}
