% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_signature_collection}
\alias{assert_signature_collection}
\alias{signature_collection}
\alias{example_signature_collection}
\title{Signature Collections}
\usage{
assert_signature_collection(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_signature_collection()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A list containing multiple data.frames, each representing a valid signature.
}
\description{
Assert object represents a signature collection

\strong{Signature Collections}: A list of signature \code{data.frames}, where each list entry represents a signature, and the name of each entry corresponds to the signature’s name.

This function returns a list of exemplar signatures with valid data following the 'sigverse' style. The list represents a collection of valid signatures.
}
\examples{
example_signature_collection()

}
