% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_signature_collection_tidy}
\alias{example_signature_collection_tidy}
\title{Exemplar collection of valid signatures}
\usage{
example_signature_collection_tidy()
}
\value{
A \code{data.frame} with 4 columns:
\describe{
\item{signature}{Name of the signature (e.g., \code{"sig1"}, \code{"sig2"})}
\item{type}{Mutation type (e.g., \code{"T>C"})}
\item{channel}{Mutation channel (e.g., \code{"A[T>C]G"})}
\item{fraction}{Normalized mutation fraction (sums to 1 per signature)}
}
}
\description{
This function returns an exemplar tidy dataframe following the 'sigverse' style.
}
\examples{
example_signature_collection_tidy()

}
