% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{sigstory_visualisation}
\alias{sigstory_visualisation}
\title{Create a sigstory Visualization Object}
\usage{
sigstory_visualisation(
  sample,
  sigclass,
  number_of_mutations,
  model_fit,
  unexplained_mutations,
  analysis_details,
  ls_signatures_minified = NULL,
  ls_exposure_dotplots = NULL,
  ls_similar_samples = NULL,
  similar_sample_disease_donut = NULL,
  umap = NULL,
  stability = NULL
)
}
\arguments{
\item{sample}{String. Sample Identifier.}

\item{sigclass}{String. The class of signature features used (e.g., SBS96, ID83, SV32, etc).}

\item{number_of_mutations}{Number. Total number of mutations in this sigclass.}

\item{model_fit}{Number. Measured as cosine similarity between observed profile vs model.}

\item{unexplained_mutations}{Number. Number of mutations not explained by signature.}

\item{analysis_details}{data.frame. A data.frame describing the details of the analysis (software versions, thresholds, reference cohorts, etc).}

\item{ls_signatures_minified}{List. A list of signatures (in model) rendered as minified ggplots. Names should be concatenations of each signature and its aetiology. If NULL, we assume no signature model could be fit. See \code{\link[sigvis:sig_visualise_minified]{sigvis::sig_visualise_minified()}}.}

\item{ls_exposure_dotplots}{List (optional). A list of signatures (in model) rendered as dotplots describing contribution size compared to a reference cohort. Default is \code{NULL}.}

\item{ls_similar_samples}{List (optional). A list of the most similar samples in the reference cohort (minified with similarity). Names should be sample ID concatenated with 'disease'. Default is \code{NULL}.}

\item{similar_sample_disease_donut}{ggplot (optional). A donut plot of the diseases of the top N similar samples. Default is \code{NULL}.}

\item{umap}{ggplot (optional). UMAP plot showing how sample clusters against the reference cohort. Default is \code{NULL}.}

\item{stability}{ggplot (optional). Signature stability plot (bootstraps). Default is \code{NULL}.}
}
\value{
An object of class \code{sigstory} containing the provided data and visualizations.
}
\description{
Constructs a \code{sigstory} object containing visualizations and data related to mutational signature analysis.
Many of the visualizations required for this object can be produced using the \code{sigvis} package.
}
\note{
Many of the visualizations required for this object can be produced using the \code{sigvis} package.
}
\examples{
\dontrun{
sigstory_obj <- sigstory_visualisation(
  sample = "Sample1",
  sigclass = "SBS96",
  number_of_mutations = 500,
  model_fit = 0.95,
  unexplained_mutations = 25,
  ls_signatures_minified = list(Sig1 = ggplot_obj1, Sig2 = ggplot_obj2),
  ls_exposure_dotplots = list(Sig1 = ggplot_obj3),
  ls_similar_samples = list(SampleA = ggplot_obj4),
  similar_sample_disease_donut = ggplot_obj5,
  umap = ggplot_obj6,
  stability = ggplot_obj7,
  analysis_details = data.frame(Detail = "Info")
)
}
}
