% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_similarity_against_cohort}
\alias{assert_similarity_against_cohort}
\alias{similarity_against_cohort}
\alias{example_similarity_against_cohort}
\title{Similarity Against Cohort Data Type}
\usage{
assert_similarity_against_cohort(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_similarity_against_cohort()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
\code{\link[=assert_similarity_against_cohort]{assert_similarity_against_cohort()}} throws an error if the assertion fails; otherwise, returns \code{TRUE} invisibly.

\code{\link[=example_similarity_against_cohort]{example_similarity_against_cohort()}} returns a \code{data.frame} representing valid similarity against cohort data.
}
\description{
Assert object represents \verb{Similarity against cohort dataset}

\strong{Similarity Against Cohort}: A \code{data.frame} that describes how similar a sample catalogue is to others in the cohort.
The data.frame should contain two columns:
\itemize{
\item \code{sample}: The sample identifier (a character or factor with no duplicates or missing values).
\item \code{cosine_similarity}: Numeric value representing the cosine similarity between the sample and another sample.
}

Note: We assume this full table describes similarity against a single sample, i.e., there should be no duplicate samples.

This function returns an exemplar \code{data.frame} representing similarity against a cohort, following the 'sigverse' style.
}
\examples{
example_similarity_against_cohort()

}
