% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/assertions.R
\name{example_umap}
\alias{example_umap}
\alias{assert_umap}
\alias{umap}
\title{Exemplar UMAP dataset}
\usage{
example_umap()

assert_umap(obj, msg = NULL, call = rlang::caller_env(), arg_name = NULL)
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
\code{\link[=example_umap]{example_umap()}} returns a \code{data.frame} representing valid UMAP dataset.

\code{\link[=assert_umap]{assert_umap()}} throws an error if the assertion fails; otherwise, returns \code{TRUE} invisibly.
}
\description{
This function returns an exemplar UMAP \code{data.frame} with valid data following the 'sigverse' style.

Assert object represents UMAP dataset

\strong{UMAP}: A \code{data.frame} representing UMAP coordinates with at least the following 3 columns:
\itemize{
\item \code{sample}: The sample identifier (a character or factor with no duplicates or missing values).
\item \code{dim1}: Numeric value representing the first UMAP dimension.
\item \code{dim2}: Numeric value representing the second UMAP dimension.
}
}
\examples{
umap <- example_umap()
assert_umap(umap)
}
