% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{valid_bootstraps}
\alias{valid_bootstraps}
\alias{example_valid_bootstrap}
\alias{example_valid_bootstrap_empty}
\title{Exemplar bootstrap with valid data}
\usage{
example_valid_bootstrap()

example_valid_bootstrap_empty()
}
\value{
A data.frame representing a valid exemplar bootstrap in the 'sigverse' format.
}
\description{
This function returns an exemplar bootstrap with valid data following the 'sigverse' style.
}
\examples{

# Return a valid bootstrap
example_valid_bootstrap()

# Return a valid but empty bootstrap (all contributions are 0)
example_valid_bootstrap_empty()
}
