% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{valid_catalogue}
\alias{valid_catalogue}
\alias{example_valid_catalogue}
\alias{example_valid_catalogue_empty}
\title{Exemplar valid catalogue}
\usage{
example_valid_catalogue()

example_valid_catalogue_empty()
}
\value{
A data.frame containing valid catalogue data.
}
\description{
This function returns a data.frame representing a valid catalogue following the 'sigverse' style.
}
\examples{

# Example of a valid catalogue
example_valid_catalogue()

# Example of a valid but empty catalogue
example_valid_catalogue_empty()

}
