% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{valid_sig}
\alias{valid_sig}
\alias{example_valid_signature}
\alias{example_valid_signature_empty}
\title{Exemplar signature and with valid data}
\usage{
example_valid_signature()

example_valid_signature_empty()
}
\value{
A data.frame representing a valid exemplar signature signature in the 'sigverse' format.
}
\description{
This function returns an exemplar signature with valid data following the 'sigverse' style.
}
\examples{

# Return a valid signature
example_valid_signature()

# Return a valid but empty signature (all fraction values are 0)
example_valid_signature_empty()
}
