% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_fraction}
\alias{compute_fraction}
\title{Convert counts into fractions of total counts}
\usage{
compute_fraction(counts)
}
\arguments{
\item{counts}{a numeric vector of counts}
}
\value{
a numeric vector of fractions
}
\description{
Convert counts into fractions of total counts, dealing with zero sum counts by setting to zero instead of Inf
}
\details{
Uses absolute value of counts for computing fractions.
}
\examples{
compute_fraction(c(10, 2, 1000, 50, 250))

}
