% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R, R/example_data.R
\name{assert_signature_collection}
\alias{assert_signature_collection}
\alias{signature_collection}
\alias{example_signature_collection}
\alias{example_signature_collection_matrix}
\title{Signature Collections}
\usage{
assert_signature_collection(
  obj,
  msg = NULL,
  call = rlang::caller_env(),
  arg_name = NULL
)

example_signature_collection()

example_signature_collection_matrix()
}
\arguments{
\item{obj}{object}

\item{msg}{error message}

\item{call}{internal paramater}

\item{arg_name}{argument}
}
\value{
A list containing multiple data.frames, each representing a valid signature.

A numeric matrix of signature fractions with:
\itemize{
\item rownames as channels,
\item colnames as signature names,
\item \code{type} attribute storing mutation type per channel.
}
}
\description{
Assert object represents a signature collection

\strong{Signature Collections}: A list of signature \code{data.frames}, where each list entry represents a signature, and the name of each entry corresponds to the signature’s name.

This function returns a list of exemplar signatures with valid data following the 'sigverse' style. The list represents a collection of valid signatures.

Returns an example of a signature collection formatted as a matrix.
}
\details{
In this matrix form:
\itemize{
\item \strong{Rows} are mutation channels.
\item \strong{Columns} are signature names.
\item \strong{Values} are normalized mutation fractions.
}

Includes an optional \code{"type"} attribute that describes the broader mutation type of each row (channel).
This format is used by functions such as \code{\link[=sig_collection_reformat_matrix_to_list]{sig_collection_reformat_matrix_to_list()}}.
}
\examples{
example_signature_collection()

example_signature_collection_matrix()

}
